added weapon

+++ Services/CharacterService/CharacterService.cs
@@ -1,3 +1,4 @@
+using System.Security.Claims;
 using AutoMapper;
 using dotnet_rpg.Data;
 using dotnet_rpg.Dtos.Character;
@@ -10,27 +11,39 @@ namespace dotnet_rpg.Services.CharacterService
   {
     private readonly IMapper _mapper;
     private readonly DataContext _context;
+    private readonly IHttpContextAccessor _httpContextAccessor;
 
-    public CharacterService(IMapper mapper, DataContext context)
+    public CharacterService(IMapper mapper, DataContext context, IHttpContextAccessor httpContextAccessor)
     {
+      _httpContextAccessor = httpContextAccessor;
       _context = context;
       _mapper = mapper;
     }
+
+    private int GetUserId() => int.Parse(_httpContextAccessor.HttpContext.User.FindFirstValue(ClaimTypes.NameIdentifier));
 
     public async Task<ServiceResponse<List<GetCharacterDto>>> AddCharacter(AddCharacterDto newCharacter)
     {
       var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
       var character = _mapper.Map<Character>(newCharacter);
+
+      character.User = await _context.Users.FirstOrDefaultAsync(user => user.Id == GetUserId());
+
       _context.Characters.Add(character);
+
       await _context.SaveChangesAsync();
-      serviceResponse.Data = await _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+
+      serviceResponse.Data = await _context.Characters
+      .Where(character => character.User.Id == GetUserId())
+      .Select(c => _mapper.Map<GetCharacterDto>(c)).ToListAsync();
+
       return serviceResponse;
     }
 
     public async Task<ServiceResponse<List<GetCharacterDto>>> GetAllCharacters()
     {
       var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
-      var dbCharacters = await _context.Characters.ToListAsync();
+      var dbCharacters = await _context.Characters.Where(character => character.User.Id == GetUserId()).ToListAsync();
       serviceResponse.Data = dbCharacters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
       return serviceResponse;
     }
@@ -38,7 +51,7 @@ namespace dotnet_rpg.Services.CharacterService
     public async Task<ServiceResponse<GetCharacterDto>> GetCharacterById(int id)
     {
       var serviceResponse = new ServiceResponse<GetCharacterDto>();
-      var dbCharacter = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id);
+      var dbCharacter = await _context.Characters.FirstOrDefaultAsync(character => character.Id == id && character.User.Id == GetUserId());
       serviceResponse.Data = _mapper.Map<GetCharacterDto>(dbCharacter);
       return serviceResponse;
     }
@@ -48,18 +61,31 @@ namespace dotnet_rpg.Services.CharacterService
       var serviceResponse = new ServiceResponse<GetCharacterDto>();
       try
       {
-        var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
-        character.Name = updatedCharacter.Name;
-        character.HitPoints = updatedCharacter.HitPoints;
-        character.Strength = updatedCharacter.Strength;
-        character.Defense = updatedCharacter.Defense;
-        character.Intelligence = updatedCharacter.Intelligence;
-        character.Class = updatedCharacter.Class;
+        //include gets related objects as well
+        var character = await _context.Characters
+        .Include(c => c.User)
+        .FirstOrDefaultAsync(c => c.Id == updatedCharacter.Id);
 
-        await _context.SaveChangesAsync();
+        if (character.User.Id == GetUserId())
+        {
+          character.Name = updatedCharacter.Name;
+          character.HitPoints = updatedCharacter.HitPoints;
+          character.Strength = updatedCharacter.Strength;
+          character.Defense = updatedCharacter.Defense;
+          character.Intelligence = updatedCharacter.Intelligence;
+          character.Class = updatedCharacter.Class;
 
-        serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
-        serviceResponse.Message = "character updated!";
+          await _context.SaveChangesAsync();
+
+          serviceResponse.Data = _mapper.Map<GetCharacterDto>(character);
+          serviceResponse.Message = "character updated!";
+        }
+        else
+        {
+          serviceResponse.Success = false;
+          serviceResponse.Message = "Character not found!";
+        }
+
       }
       catch (Exception ex)
       {
@@ -75,11 +101,22 @@ namespace dotnet_rpg.Services.CharacterService
       var serviceResponse = new ServiceResponse<List<GetCharacterDto>>();
       try
       {
-        var character = await _context.Characters.FirstAsync(c => c.Id == id);
-        _context.Characters.Remove(character);
-        await _context.SaveChangesAsync();
-        serviceResponse.Data = _context.Characters.Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
-        serviceResponse.Message = "character deleted!";
+        var character = await _context.Characters.FirstOrDefaultAsync(c => c.Id == id && c.User.Id == GetUserId());
+        if (character != null)
+        {
+          _context.Characters.Remove(character);
+          await _context.SaveChangesAsync();
+
+          serviceResponse.Data = _context.Characters
+          .Where(c => c.User.Id == GetUserId())
+          .Select(c => _mapper.Map<GetCharacterDto>(c)).ToList();
+          serviceResponse.Message = "character deleted!";
+        }
+        else
+        {
+          serviceResponse.Success = false;
+          serviceResponse.Message = "Character not found!";
+        }
       }
       catch (Exception ex)
       {
